+++ index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Document</title>
   <link rel="stylesheet" href="index.css">
-  <script src="modules/index.js" defer type="module"></script>
+  <script src="./modules/index.js" defer type="module"></script>
 </head>
 <body>
   <header>
